Creating comments and sub comments tested and working in t2 and t3

+++ BlazorWASM-AppLogic/Application/DAOsInterfaces/ILikeDao.cs
@@ -5,7 +5,7 @@ namespace Application.DAOsInterfaces;
 
 public interface ILikeDao
 {
-    Task<int> CountLikesAsync(int postId);
+    Task<int> CountLikesAsync(long postId);
     Task<Like> LikePostAsync(LikePostDto likeDto);
     Task UnLikePostAsync(UnLikePostDto unLikePostDto);
     Task<bool> IsPostLikedAsync(IsPostLikedDto isPostLikedDto);
