+++ BlazorWASM-AppLogic/Domain/DTOs/IsPostLikedDto.cs
@@ -2,10 +2,10 @@
 
 public class IsPostLikedDto
 {
-    public int postId { get; }
+    public long postId { get; }
     public long userId { get; }
 
-    public IsPostLikedDto(int postId, long userId)
+    public IsPostLikedDto(long postId, long userId)
     {
         this.postId = postId;
         this.userId = userId;
