+++ BlazorWASM-AppLogic/Domain/DTOs/SearchPostParametersDto.cs
@@ -2,11 +2,11 @@ namespace Domain.DTOs;
 
 public class SearchPostParametersDto
 {
-	public int? Id { get; }
-	public int? UserId { get;}
+	public long? Id { get; }
+	public long? UserId { get;}
 	public string? TitleContains { get;}
 
-	public SearchPostParametersDto(int? id = null, int? userId = null, string? titleContains = null)
+	public SearchPostParametersDto(long? id = null, long? userId = null, string? titleContains = null)
 	{
 		Id = id;
 		UserId = userId;
